Convert relative src references to absolute paths
When including files from a sub-site or a sub-folder within the same
site, Markbind only automatically resolves src references for the
include and panel tags. This means that references to any other
resources (such as anchor#href, img#src and pic#src) do not get
resolved and converted.

When these markdown files containing relative resource references are
included by another file in a different directory, the links remain
unchanged. From the perspective of the outer document, the references
become broken and files become not found.

To resolve this problem, let's convert all relative resource
references to absolute ones. This allows the references to work
consistently across all instances of reuse.

This conversion process adds on to the existing preprocess step and is
similar to how we currently resolve src references in include and
panel tags.